# BUG/PERF: DataFrame.to_records inconsistent dtypes for MultiIndex